requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pymongo==4.6.0
psycopg2-binary
python-dotenv